+++ source/UFCGuideAction/js/GuideActionManager.js
@@ -1,0 +1,17 @@
+function GuideActionManager() {}
+
+GuideActionManager.initialize = function () {
+  this._guideAction = new GuideAction();
+};
+
+Object.defineProperty(GuideActionManager, "getGuideAction", {
+  get: function () {
+    return this._guideAction;
+  },
+});
+
+Object.defineProperty(GuideActionManager, "getGuideActionGraphics", {
+  get: function () {
+    return this.getGuideAction.getGraphics();
+  },
+});