+++ source/UFCTowerDefense/js/ufcTowerSpawnData.js
@@ -3,6 +3,7 @@ const ufcTowerSpawnData = function () {
   this.initialize(...arguments);
 };
 
+ufcTowerSpawnData.prototype.constructor = ufcTowerSpawnData;
 ufcTowerSpawnData.prototype.initialize = function (data) {
   this._x = data._x;
   this._y = data._y;